mpu: move to platform as not all plaforms use ARM's mpu
- Kinetis MCU from FSL uses a different MPU than ARMs

Fix#78
